refactor(ui): split UserPopOverCard and fetch team popover via react-query
Address review feedback and modularize the popover card:
- Replace the hand-rolled team promise cache with a useTeamPopoverData
  react-query hook (5 min staleTime, shared QueryClient gc) so cached
  teams render without a loader flash and stale entries revalidate.
- Use singular/plural user label based on team user count.
- Extract UserTeams, UserRoles, PopoverContent, PopoverTitle,
  TeamPopoverContent and TeamPopoverTitle into dedicated components
  with individual unit tests; UserPopOverCard keeps only the popover
  shell. Titles move from antd Button/Space to core-components Button.
- Fix missing hook dependencies in PopoverContent effect/callback.
