+++ openmetadata-ui/src/main/resources/ui/src/hooks/team/useTeamPopoverData.ts
@@ -1,0 +1,32 @@
+/*
+ *  Copyright 2026 Collate.
+ *  Licensed under the Apache License, Version 2.0 (the "License");
+ *  you may not use this file except in compliance with the License.
+ *  You may obtain a copy of the License at
+ *  http://www.apache.org/licenses/LICENSE-2.0
+ *  Unless required by applicable law or agreed to in writing, software
+ *  distributed under the License is distributed on an "AS IS" BASIS,
+ *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *  See the License for the specific language governing permissions and
+ *  limitations under the License.
+ */
+
+import { useQuery } from '@tanstack/react-query';
+import { TabSpecificField } from '../../enums/entity.enum';
+import { getTeamByName } from '../../rest/teamsAPI';
+
+const TEAM_POPOVER_STALE_TIME_MS = 5 * 60 * 1000;
+
+export const useTeamPopoverData = (teamName: string) => {
+  const { data: team, isLoading: loading } = useQuery({
+    queryKey: ['team-popover', teamName],
+    queryFn: () =>
+      getTeamByName(teamName, {
+        fields: [TabSpecificField.PARENTS, TabSpecificField.USER_COUNT],
+      }),
+    enabled: Boolean(teamName),
+    staleTime: TEAM_POPOVER_STALE_TIME_MS,
+  });
+
+  return { team, loading };
+};
